Moving some of the more corner case feature tests into worker service tests
